added an alert to warn when deleting a photo

+++ DatingApp-SPA/src/app/members/photo-edit/photo-edit.component.ts
@@ -70,12 +70,14 @@ export class PhotoEditComponent implements OnInit {
   }
 
   onDeletePhoto(photo: Photo) {
-    this.userService.deletePhoto(+this.authService.decodedToken.nameid, photo.id).subscribe( responseData => {
-      this.photos.splice(this.photos.findIndex(p => p.id == photo.id), 1);
-      this.alertifyService.success('photo deleted successfully');
-      },
-      error => {
-      this.alertifyService.errorMessage(error);
-      })
+    this.alertifyService.confirm('are you sure you would like to delete this photo', () => {
+      this.userService.deletePhoto(+this.authService.decodedToken.nameid, photo.id).subscribe( responseData => {
+          this.photos.splice(this.photos.findIndex(p => p.id === photo.id), 1);
+          this.alertifyService.success('photo deleted successfully');
+        },
+        error => {
+          this.alertifyService.errorMessage(error);
+        });
+    });
   }
 }
